pyproject.toml
[project]
name = "goit-ds-hw-03"
version = "0.1.0"
description = ""
authors = [
    {name = "mykyta",email = "nikita.slipachuk@gmail.com"}
]
readme = "README.md"
requires-python = "3.12.2"
dependencies = [
    "pymongo (>=4.15.3,<5.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "beautifulsoup4 (>=4.14.2,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
